+++ packages/recommend/src/types/TrendingFacetHit.ts
@@ -1,0 +1,5 @@
+export type TrendingFacetHit<TObject> = {
+  readonly _score: number;
+  readonly facetName: string;
+  readonly facetValue: TObject;
+};
