+++ packages/recommend/src/types/TrendingFacetsResponse.ts
@@ -1,0 +1,7 @@
+import { SearchResponse } from '@algolia/client-search';
+
+import { TrendingFacetHit } from './TrendingFacetHit';
+
+export type TrendingFacetsResponse<TObject> = Omit<SearchResponse<TObject>, 'hits'> & {
+  readonly hits: ReadonlyArray<TrendingFacetHit<TObject>>;
+};
