+++ packages/recommend/src/types/index.ts
@@ -15,3 +15,5 @@ export * from './TrendingItemsQuery';
 export * from './TrendingQuery';
 export * from './WithRecommendMethods';
 export * from './LookingSimilarQuery';
+export * from './TrendingFacetHit';
+export * from './TrendingFacetsResponse';
